# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Backend Developer Node.js Study Lab API
  version: 1.1.0
  description: >-
    Complete OpenAPI reference for the backend study cases.
    Includes request body schemas, request examples, and response examples
    for both success and common error outcomes.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /cases/07-auth/login:
    post:
      tags: [Foundations]
      summary: Sign a JWT for a user role
      operationId: loginCase07
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLogin"
            examples:
              developerLogin:
                value:
                  username: dev-user
                  role: developer
              adminLogin:
                value:
                  username: platform-admin
                  role: admin
      responses:
        "200":
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                success:
                  value:
                    case: 07-auth
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.sample
                    user:
                      username: dev-user
                      role: developer
        "400":
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidRole:
                  value:
                    error:
                      fieldErrors:
                        role:
                          - Invalid enum value. Expected 'admin' | 'developer' | 'viewer'
                      formErrors: []
...
  /cases/11-ai-ml/summarize:
    post:
      tags: [Product and Observability]
      summary: Summarize text using provider or local fallback
      operationId: summarizeCase11Text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SummarizeRequest"
            examples:
              summarize:
                value:
                  text: Modern backend systems need resilient integrations, clear observability, and predictable fallback strategies when providers become unavailable.
      responses:
        "200":
          description: Summary generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                localFallback:
                  value:
                    case: 11-ai-ml
                    provider: local-fallback
                    summary: Modern backend systems need resilient integrations, clear observability, and predictable fallback strategies when providers become unavailable.
        "400":
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                tooShort:
                  value:
                    error:
                      fieldErrors:
                        text:
                          - String must contain at least 20 character(s)
                      formErrors: []
        "502":
          description: Provider failure with fallback response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                providerFailure:
                  value:
                    case: 11-ai-ml
                    error: AI provider call failed
                    fallbackSummary: Modern backend systems need resilient integrations ...
...
  /cases/16-resilience/invoke:
    post:
      tags: [Reliability and Domain]
      summary: Run dependency call with retries, timeout, circuit breaker, and bulkhead
      operationId: invokeCase16Resilience
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResilienceInvoke"
            examples:
              successScenario:
                value:
                  dependency: payment-gateway
                  scenario: success
                  timeoutMs: 180
              flakyScenario:
                value:
                  dependency: payment-gateway
                  scenario: flaky
                  timeoutMs: 180
      responses:
        "200":
          description: Invocation succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                success:
                  value:
                    case: 16-resilience
                    dependency: payment-gateway
                    scenario: flaky
                    attemptsUsed: 2
                    elapsedMs: 49
                    status: success
                    result:
                      dependency: payment-gateway
                      scenario: flaky
                      attempt: 2
                      ok: true
                    circuit:
                      dependency: payment-gateway
                      status: closed
                      consecutiveFailures: 0
                      openedAtMs: null
                      inflight: 0
                      calls: 3
                      successes: 2
                      failures: 1
                      timeouts: 0
                      rejected: 0
                      lastError: null
                      updatedAt: 2026-04-11T09:40:10.220Z
                    policy:
                      maxAttempts: 3
                      timeoutMs: 180
                      failureThreshold: 3
                      openWindowMs: 8000
                      bulkheadLimit: 4
        "400":
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                validationError:
                  value:
                    case: 16-resilience
                    error:
                      fieldErrors:
                        dependency:
                          - String must contain at least 2 character(s)
                      formErrors: []
        "429":
          description: Rejected by bulkhead concurrency limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                bulkheadReject:
                  value:
                    case: 16-resilience
                    dependency: payment-gateway
                    status: rejected_by_bulkhead
                    circuit:
                      dependency: payment-gateway
                      inflight: 4
                    policy:
                      bulkheadLimit: 4
        "502":
          description: Failed after retries (non-timeout)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                downstreamFailure:
                  value:
                    case: 16-resilience
                    dependency: payment-gateway
                    scenario: fail
                    attemptsUsed: 3
                    elapsedMs: 104
                    status: failed_after_retries
                    error: downstream_failure
        "503":
          description: Rejected by open circuit breaker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                openCircuit:
                  value:
                    case: 16-resilience
                    dependency: payment-gateway
                    status: rejected_by_circuit
                    retryAfterMs: 5230
        "504":
          description: Failed after retries due to timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                timeoutFailure:
                  value:
                    case: 16-resilience
                    dependency: payment-gateway
                    scenario: timeout
                    attemptsUsed: 3
                    elapsedMs: 603
                    status: failed_after_retries
                    error: request_timeout
...
  /cases/17-event-sourcing-cqrs/accounts:
    post:
      tags: [Reliability and Domain]
      summary: Open account and append first event
      operationId: openCase17Account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAccountRequest"
            examples:
              openAccount:
                value:
                  accountId: acct-test-1
                  initialBalance: 250
                  commandId: cmd-open-acct-test-1
      responses:
        "201":
          description: Account opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                created:
                  value:
                    case: 17-event-sourcing-cqrs
                    status: accepted
                    event:
                      sequence: 1
                      accountId: acct-test-1
                      commandId: cmd-open-acct-test-1
                      type: account-opened
                      amount: 250
                      occurredAt: 2026-04-11T09:45:00.000Z
                    projection:
                      accountId: acct-test-1
                      balance: 250
                      version: 1
                      updatedAt: 2026-04-11T09:45:00.000Z
        "200":
          description: Duplicate command accepted idempotently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                duplicate:
                  value:
                    case: 17-event-sourcing-cqrs
                    status: duplicate_command
                    commandId: cmd-open-acct-test-1
                    eventSequence: 1
                    projection:
                      accountId: acct-test-1
                      balance: 250
                      version: 1
                      updatedAt: 2026-04-11T09:45:00.000Z
        "400":
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                validationError:
                  value:
                    case: 17-event-sourcing-cqrs
                    error:
                      fieldErrors:
                        initialBalance:
                          - Number must be greater than or equal to 0
                      formErrors: []
        "409":
          description: Account already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                alreadyExists:
                  value:
                    case: 17-event-sourcing-cqrs
                    error: "account already exists: acct-test-1"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
      required: [error]
...
    AuthLogin:
      type: object
      properties:
        username:
          type: string
          minLength: 2
        role:
          type: string
          enum: [admin, developer, viewer]
          default: developer
      required: [username]
...
    SummarizeRequest:
      type: object
      properties:
        text:
          type: string
          minLength: 20
      required: [text]
    ResilienceInvoke:
      type: object
      properties:
        dependency:
          type: string
          minLength: 2
          maxLength: 64
          default: payment-gateway
        scenario:
          type: string
          enum: [success, fail, timeout, flaky]
          default: success
        timeoutMs:
          type: integer
          minimum: 60
          maximum: 3000
      required: [dependency]
    OpenAccountRequest:
      type: object
      properties:
        accountId:
          type: string
        initialBalance:
          type: number
          minimum: 0
          maximum: 1000000
        commandId:
          type: string
      required: [initialBalance]
...